backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
python-dotenv==1.0.1
python-jose==3.3.0
bcrypt==3.2.2
httpx==0.27.0
